requirements.txt
django
djangorestframework
pillow
celery
django-deepzoom